+++ CHANGELOG.md
@@ -1,4 +1,14 @@
+<!--
+SPDX-FileCopyrightText: 2025 DAASI International GmbH <info@daasi.de>
+SPDX-License-Identifier: Apache-2.0
+-->
+
 # CHANGELOG edumfa-radius
+
+## v0.0.2
+
+- This release contains no functional changes.
+- IMPROVEMENT: add installation instructions
 
 ## v0.0.1